+++ client/src/components/NeedList/index.jsx
@@ -1,12 +1,24 @@
+import { useState } from "react";
 import { Link } from "react-router-dom";
+import AuthService from "../../utils/auth";
 
 const NeedList = ({ needs, title, showTitle = true, showFirstName = true }) => {
+  const isAuthenticated = AuthService.loggedIn();
+  const [message, setMessage] = useState("");
+
+  const messageHandler = () => {
+    setMessage("Please login or sign-up if you would like to participate!");
+  }
+
   if (!needs.length) {
     return <h3 style={{textAlign: "center", marginTop: "11rem", marginBottom: "11rem"}}>No Community Needs Yet</h3>;
   }
 
   return (
     <div>
+      <div>
+        <p>{message}</p>
+      </div>
       {showTitle && <h3>{title}</h3>}
       {needs &&
         needs.map((need) => (
@@ -37,12 +49,21 @@ const NeedList = ({ needs, title, showTitle = true, showFirstName = true }) => {
               <p>{need.needText}</p>
               <p>Project date: {need.needDate}</p>
             </div>
-            <Link
-              className="btn btn-primary btn-block btn-squared"
-              to={`/needs/${need._id}`}
-            >
-              Participate in this community project.
-            </Link>
+            { isAuthenticated ? (
+              <Link
+                className="btn btn-primary btn-block btn-squared"
+                to={`/needs/${need._id}`}
+              >
+                Participate in this community project.
+              </Link>
+            ) : (
+              <Link
+                className="btn btn-primary btn-block btn-squared"
+                onClick={messageHandler}
+              >
+                Participate in this community project.
+              </Link>
+            )}
           </div>
         ))}
     </div>
